Found a bug dealing with sync limits
limit happens after aggregate so updates are effectively skipped if they
have a later change in a sequence that appears after the limit.

ex sequence=100 section created
   sequence=150 section updated

get syncs sequence 99 limit 3

this would not get the section created with 100 because its sequence was
set to 150 from the custom aggregate.

a possible solution is to either limit before the aggregate or try to
figure out beforehand how the sequence of the last limited record after
folding